Implement audio node connections in Rust
This approach ditches the whole audioNode.connect scheme provided
by WebAudio, and instead connects all the nodes inside the WASM
part. This should improve the performance quite a bit, as it requires
no back-and-forth with the WebAudio engine and instead calculates
everything in a single .process call of a WebAudioWorklet.